tests/wpt/tests/css/css-color-hdr/interpolation.html

<!DOCTYPE html>
<title>Interpolation of dynamic-range-limit</title>
<link rel="help" href="https://drafts.csswg.org/css-color-hdr/#the-dynamic-range-limit-property">
<meta name="assert" content="Interpolation of dynamic-range-limit">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
  <template id="target-template">T</template>
</body>

<script>
test_interpolation({
  property: "dynamic-range-limit",
  from: "no-limit",
  to: "standard",
}, [
  {at: 0.0, expect: "no-limit"},
  {at: 0.25, expect: "dynamic-range-limit-mix(standard 25%, no-limit 75%)"},
  {at: 0.75, expect: "dynamic-range-limit-mix(standard 75%, no-limit 25%)"},
  {at: 1.0, expect: "standard"},
]);

test_interpolation({
  property: "dynamic-range-limit",
  from: "constrained",
  to: "standard",
}, [
  {at: 0.0, expect: "constrained"},
  {at: 0.5, expect: "dynamic-range-limit-mix(standard 50%, constrained 50%)"},
  {at: 1.0, expect: "standard"},
]);

test_interpolation({
  property: "dynamic-range-limit",
  from: "dynamic-range-limit-mix(standard 50%, no-limit 50%)",
  to: "standard",
}, [
  {at: 0.0, expect: "dynamic-range-limit-mix(standard 50%, no-limit 50%)"},
  {at: 0.75, expect: "dynamic-range-limit-mix(standard 87.5%, no-limit 12.5%)"},
  {at: 1.0, expect: "standard"},
]);

test_interpolation({
  property: "dynamic-range-limit",
  from: "dynamic-range-limit-mix(standard 90%, no-limit 10%)",
  to: "dynamic-range-limit-mix(standard 10%, no-limit 90%)",
}, [
  {at: 0.0, expect: "dynamic-range-limit-mix(standard 90%, no-limit 10%)"},
  {at: 0.5, expect: "dynamic-range-limit-mix(standard 50%, no-limit 50%)"},
  {at: 1.0, expect: "dynamic-range-limit-mix(standard 10%, no-limit 90%)"},
]);

test_interpolation({
  property: "dynamic-range-limit",
  from: "dynamic-range-limit-mix(constrained 90%, standard 10%)",
  to: "dynamic-range-limit-mix(no-limit 10%, standard 90%)",
}, [
  {at: 0.0, expect: "dynamic-range-limit-mix(standard 10%, constrained 90%)"},
  {at: 0.5, expect: "dynamic-range-limit-mix(standard 50%, constrained 45%, no-limit 5%)"},
  {at: 1.0, expect: "dynamic-range-limit-mix(standard 90%, no-limit 10%)"},
]);
</script>
