LayoutTests/imported/w3c/web-platform-tests/css/css-color-hdr/interpolation-expected.txt


FAIL CSS Transitions: property <dynamic-range-limit> from [no-limit] to [standard] at (0) should be [no-limit] assert_equals: expected "no - limit " but got "standard "
FAIL CSS Transitions: property <dynamic-range-limit> from [no-limit] to [standard] at (0.25) should be [dynamic-range-limit-mix(standard 25%, no-limit 75%)] assert_equals: expected "dynamic - range - limit - mix ( standard 25 % , no - limit 75 % ) " but got "standard "
FAIL CSS Transitions: property <dynamic-range-limit> from [no-limit] to [standard] at (0.75) should be [dynamic-range-limit-mix(standard 75%, no-limit 25%)] assert_equals: expected "dynamic - range - limit - mix ( standard 75 % , no - limit 25 % ) " but got "standard "
PASS CSS Transitions: property <dynamic-range-limit> from [no-limit] to [standard] at (1) should be [standard]
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [no-limit] to [standard] at (0) should be [no-limit] assert_equals: expected "no - limit " but got "standard "
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [no-limit] to [standard] at (0.25) should be [dynamic-range-limit-mix(standard 25%, no-limit 75%)] assert_equals: expected "dynamic - range - limit - mix ( standard 25 % , no - limit 75 % ) " but got "standard "
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [no-limit] to [standard] at (0.75) should be [dynamic-range-limit-mix(standard 75%, no-limit 25%)] assert_equals: expected "dynamic - range - limit - mix ( standard 75 % , no - limit 25 % ) " but got "standard "
PASS CSS Transitions with transition: all: property <dynamic-range-limit> from [no-limit] to [standard] at (1) should be [standard]
PASS CSS Animations: property <dynamic-range-limit> from [no-limit] to [standard] at (0) should be [no-limit]
FAIL CSS Animations: property <dynamic-range-limit> from [no-limit] to [standard] at (0.25) should be [dynamic-range-limit-mix(standard 25%, no-limit 75%)] assert_equals: expected "dynamic - range - limit - mix ( standard 25 % , no - limit 75 % ) " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [no-limit] to [standard] at (0.75) should be [dynamic-range-limit-mix(standard 75%, no-limit 25%)] assert_equals: expected "dynamic - range - limit - mix ( standard 75 % , no - limit 25 % ) " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [no-limit] to [standard] at (1) should be [standard] assert_equals: expected "standard " but got "no - limit "
PASS Web Animations: property <dynamic-range-limit> from [no-limit] to [standard] at (0) should be [no-limit]
FAIL Web Animations: property <dynamic-range-limit> from [no-limit] to [standard] at (0.25) should be [dynamic-range-limit-mix(standard 25%, no-limit 75%)] assert_equals: expected "dynamic - range - limit - mix ( standard 25 % , no - limit 75 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [no-limit] to [standard] at (0.75) should be [dynamic-range-limit-mix(standard 75%, no-limit 25%)] assert_equals: expected "dynamic - range - limit - mix ( standard 75 % , no - limit 25 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [no-limit] to [standard] at (1) should be [standard] assert_equals: expected "standard " but got "no - limit "
FAIL CSS Transitions: property <dynamic-range-limit> from [constrained] to [standard] at (0) should be [constrained] assert_equals: expected "constrained " but got "standard "
FAIL CSS Transitions: property <dynamic-range-limit> from [constrained] to [standard] at (0.5) should be [dynamic-range-limit-mix(standard 50%, constrained 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , constrained 50 % ) " but got "standard "
PASS CSS Transitions: property <dynamic-range-limit> from [constrained] to [standard] at (1) should be [standard]
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [constrained] to [standard] at (0) should be [constrained] assert_equals: expected "constrained " but got "standard "
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [constrained] to [standard] at (0.5) should be [dynamic-range-limit-mix(standard 50%, constrained 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , constrained 50 % ) " but got "standard "
PASS CSS Transitions with transition: all: property <dynamic-range-limit> from [constrained] to [standard] at (1) should be [standard]
FAIL CSS Animations: property <dynamic-range-limit> from [constrained] to [standard] at (0) should be [constrained] assert_equals: expected "constrained " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [constrained] to [standard] at (0.5) should be [dynamic-range-limit-mix(standard 50%, constrained 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , constrained 50 % ) " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [constrained] to [standard] at (1) should be [standard] assert_equals: expected "standard " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [constrained] to [standard] at (0) should be [constrained] assert_equals: expected "constrained " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [constrained] to [standard] at (0.5) should be [dynamic-range-limit-mix(standard 50%, constrained 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , constrained 50 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [constrained] to [standard] at (1) should be [standard] assert_equals: expected "standard " but got "no - limit "
FAIL CSS Transitions: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (0) should be [dynamic-range-limit-mix(standard 50%, no-limit 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , no - limit 50 % ) " but got "standard "
FAIL CSS Transitions: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (0.75) should be [dynamic-range-limit-mix(standard 87.5%, no-limit 12.5%)] assert_equals: expected "dynamic - range - limit - mix ( standard 87.5 % , no - limit 12.5 % ) " but got "standard "
PASS CSS Transitions: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (1) should be [standard]
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (0) should be [dynamic-range-limit-mix(standard 50%, no-limit 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , no - limit 50 % ) " but got "standard "
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (0.75) should be [dynamic-range-limit-mix(standard 87.5%, no-limit 12.5%)] assert_equals: expected "dynamic - range - limit - mix ( standard 87.5 % , no - limit 12.5 % ) " but got "standard "
PASS CSS Transitions with transition: all: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (1) should be [standard]
FAIL CSS Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (0) should be [dynamic-range-limit-mix(standard 50%, no-limit 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , no - limit 50 % ) " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (0.75) should be [dynamic-range-limit-mix(standard 87.5%, no-limit 12.5%)] assert_equals: expected "dynamic - range - limit - mix ( standard 87.5 % , no - limit 12.5 % ) " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (1) should be [standard] assert_equals: expected "standard " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (0) should be [dynamic-range-limit-mix(standard 50%, no-limit 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , no - limit 50 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (0.75) should be [dynamic-range-limit-mix(standard 87.5%, no-limit 12.5%)] assert_equals: expected "dynamic - range - limit - mix ( standard 87.5 % , no - limit 12.5 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 50%, no-limit 50%)] to [standard] at (1) should be [standard] assert_equals: expected "standard " but got "no - limit "
FAIL CSS Transitions: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (0) should be [dynamic-range-limit-mix(standard 90%, no-limit 10%)] assert_equals: expected "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) " but got "dynamic - range - limit - mix ( standard 10 % , no - limit 90 % ) "
FAIL CSS Transitions: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (0.5) should be [dynamic-range-limit-mix(standard 50%, no-limit 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , no - limit 50 % ) " but got "dynamic - range - limit - mix ( standard 10 % , no - limit 90 % ) "
PASS CSS Transitions: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (1) should be [dynamic-range-limit-mix(standard 10%, no-limit 90%)]
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (0) should be [dynamic-range-limit-mix(standard 90%, no-limit 10%)] assert_equals: expected "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) " but got "dynamic - range - limit - mix ( standard 10 % , no - limit 90 % ) "
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (0.5) should be [dynamic-range-limit-mix(standard 50%, no-limit 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , no - limit 50 % ) " but got "dynamic - range - limit - mix ( standard 10 % , no - limit 90 % ) "
PASS CSS Transitions with transition: all: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (1) should be [dynamic-range-limit-mix(standard 10%, no-limit 90%)]
FAIL CSS Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (0) should be [dynamic-range-limit-mix(standard 90%, no-limit 10%)] assert_equals: expected "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (0.5) should be [dynamic-range-limit-mix(standard 50%, no-limit 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , no - limit 50 % ) " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (1) should be [dynamic-range-limit-mix(standard 10%, no-limit 90%)] assert_equals: expected "dynamic - range - limit - mix ( standard 10 % , no - limit 90 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (0) should be [dynamic-range-limit-mix(standard 90%, no-limit 10%)] assert_equals: expected "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (0.5) should be [dynamic-range-limit-mix(standard 50%, no-limit 50%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , no - limit 50 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(standard 90%, no-limit 10%)] to [dynamic-range-limit-mix(standard 10%, no-limit 90%)] at (1) should be [dynamic-range-limit-mix(standard 10%, no-limit 90%)] assert_equals: expected "dynamic - range - limit - mix ( standard 10 % , no - limit 90 % ) " but got "no - limit "
FAIL CSS Transitions: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (0) should be [dynamic-range-limit-mix(standard 10%, constrained 90%)] assert_equals: expected "dynamic - range - limit - mix ( standard 10 % , constrained 90 % ) " but got "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) "
FAIL CSS Transitions: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (0.5) should be [dynamic-range-limit-mix(standard 50%, constrained 45%, no-limit 5%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , constrained 45 % , no - limit 5 % ) " but got "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) "
PASS CSS Transitions: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (1) should be [dynamic-range-limit-mix(standard 90%, no-limit 10%)]
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (0) should be [dynamic-range-limit-mix(standard 10%, constrained 90%)] assert_equals: expected "dynamic - range - limit - mix ( standard 10 % , constrained 90 % ) " but got "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) "
FAIL CSS Transitions with transition: all: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (0.5) should be [dynamic-range-limit-mix(standard 50%, constrained 45%, no-limit 5%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , constrained 45 % , no - limit 5 % ) " but got "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) "
PASS CSS Transitions with transition: all: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (1) should be [dynamic-range-limit-mix(standard 90%, no-limit 10%)]
FAIL CSS Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (0) should be [dynamic-range-limit-mix(standard 10%, constrained 90%)] assert_equals: expected "dynamic - range - limit - mix ( standard 10 % , constrained 90 % ) " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (0.5) should be [dynamic-range-limit-mix(standard 50%, constrained 45%, no-limit 5%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , constrained 45 % , no - limit 5 % ) " but got "no - limit "
FAIL CSS Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (1) should be [dynamic-range-limit-mix(standard 90%, no-limit 10%)] assert_equals: expected "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (0) should be [dynamic-range-limit-mix(standard 10%, constrained 90%)] assert_equals: expected "dynamic - range - limit - mix ( standard 10 % , constrained 90 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (0.5) should be [dynamic-range-limit-mix(standard 50%, constrained 45%, no-limit 5%)] assert_equals: expected "dynamic - range - limit - mix ( standard 50 % , constrained 45 % , no - limit 5 % ) " but got "no - limit "
FAIL Web Animations: property <dynamic-range-limit> from [dynamic-range-limit-mix(constrained 90%, standard 10%)] to [dynamic-range-limit-mix(no-limit 10%, standard 90%)] at (1) should be [dynamic-range-limit-mix(standard 90%, no-limit 10%)] assert_equals: expected "dynamic - range - limit - mix ( standard 90 % , no - limit 10 % ) " but got "no - limit "
